requirements.txt
fastapi
uvicorn[standard]
websockets
dotenv
requests
ffmpeg-python
langchain
langchain_core
langchain_google_genai
beautifulsoup4
openai-whisper
sentence_transformers
numpy
httpx
duckduckgo_search
audio-extract
ddgs
pymongo
bcrypt
python-jose[cryptography]
pydantic[email]